Refactor libcall registry into dedicated modules
