+++ src/libcall.h
@@ -7,28 +7,4 @@
 
 #pragma once
 
-#include <stdint.h>
-#include <stdbool.h>
-
-#include "interpret.h"
-
-typedef struct {
-  const char *libname;
-  const char *callname;
-  int8_t lib_index;
-  int8_t call_index;
-  uint8_t args;
-  OP_t func;
-} LIBCALL_t;
-
-extern const LIBCALL_t libcalls[];
-
-bool libcall_lookup_token(const char *libname, const char *callname, uint8_t *token, uint8_t *args);
-bool libcall_token_arg_count(uint8_t token, uint8_t *args);
-bool libcall_init_registry(void);
-void libcall_free_registry(void);
-void libcall_reset_registry_for_tests(void);
-bool libcall_validate_registry(void);
-bool libcall_registry_self_check(const LIBCALL_t *calls, bool fail_fast);
-bool libcall_names_unique(const LIBCALL_t *calls);
-OP_t libcall_func_token(uint8_t token);
+#include "libcall_registry.h"
